requirements.txt
aiohttp
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bilibili-api-dev==16.2.0b1
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
lxml==4.9.4
multidict==6.0.4
openai==1.8.0
Pillow==10.1.0
pyasn1==0.5.1
pycryptodomex==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
qrcode-terminal==0.8
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4
